requirements.txt
# Core bot framework
aiogram>=3.0.0

# AI and image processing
openai>=1.0.0
Pillow>=9.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Async HTTP client
aiohttp>=3.8.0

# Data handling
pydantic>=2.0.0